classifierNet/requirements.txt
numpy
Pillow
tensorflow>=2.15
scikit-learn
matplotlib 